[Pal/Linux-SGX] Fix and refactor untrusted API code
Problems fixed:
* handle_ecall incorrectly checked ms_sec_info argument (missing address verification and potential
  TOCTOU problem).
* init_xsave_size argument was taken from an untrusted source, while it can be retrieved from the
  CPU directly.
* A lot of ocalls used wrong types for sizes.
* `noinline` hack for sgx_copy_to_enclave and sgx_copy_ptr_to_enclave is not needed anymore as we
  use READ_ONCE and WRITE_ONCE now.

+++ Pal/src/host/Linux-SGX/enclave_ocalls.h
@@ -46,7 +46,7 @@ int ocall_ftruncate (int fd, uint64_t length);
 
 int ocall_mkdir (const char *pathname, unsigned short mode);
 
-int ocall_getdents (int fd, struct linux_dirent64 *dirp, unsigned int size);
+int ocall_getdents(int fd, struct linux_dirent64* dirp, size_t size);
 
 int ocall_listen(int domain, int type, int protocol, int ipv6_v6only, struct sockaddr* addr,
                  size_t* addrlen, struct sockopt* sockopt);
